requirements.txt
# Data collection
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
APScheduler==3.10.4

# Data processing
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
xgboost==2.0.3
catboost==1.2.7

# NLP
nltk==3.8.1
sentence-transformers==2.6.1

# Frontend & API
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2

# Utils
python-dotenv==1.0.1
joblib==1.3.2
